Weaver [Command] processing moved into NetworkBehaviourCommandProcessor
