+++ Mirror/Weaver/Mirror.Weaver.csproj
@@ -68,6 +68,7 @@
     <Compile Include="AssemblyInfo.cs" />
     <Compile Include="MessageClassProcessor.cs" />
     <Compile Include="MonoBehaviourProcessor.cs" />
+    <Compile Include="NetworkBehaviourCommandProcessor.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="SyncListStructProcessor.cs" />
     <Compile Include="NetworkBehaviourProcessor.cs" />
