+++ Mirror/Weaver/NetworkBehaviourProcessor.cs
@@ -28,7 +28,6 @@ namespace Mirror.Weaver
         const int k_SyncVarLimit = 64; // ulong = 64 bytes
         readonly TypeDefinition m_td;
 
-        const string k_CmdPrefix = "InvokeCmd";
         const string k_RpcPrefix = "InvokeRpc";
         const string k_TargetRpcPrefix = "InvokeTargetRpc";
 
@@ -73,7 +72,7 @@ namespace Mirror.Weaver
             Weaver.DLog(m_td, "Process Done");
         }
 
-        static void WriteClientActiveCheck(ILProcessor worker, string mdName, Instruction label, string errString)
+        public static void WriteClientActiveCheck(ILProcessor worker, string mdName, Instruction label, string errString)
         {
             // client active check
             worker.Append(worker.Create(OpCodes.Call, Weaver.NetworkClientGetActive));
@@ -85,7 +84,7 @@ namespace Mirror.Weaver
             worker.Append(label);
         }
 
-        static void WriteServerActiveCheck(ILProcessor worker, string mdName, Instruction label, string errString)
+        public static void WriteServerActiveCheck(ILProcessor worker, string mdName, Instruction label, string errString)
         {
             // server active check
             worker.Append(worker.Create(OpCodes.Call, Weaver.NetworkServerGetActive));
@@ -97,20 +96,20 @@ namespace Mirror.Weaver
             worker.Append(label);
         }
 
-        static void WriteSetupLocals(ILProcessor worker)
+        public static void WriteSetupLocals(ILProcessor worker)
         {
             worker.Body.InitLocals = true;
             worker.Body.Variables.Add(new VariableDefinition(Weaver.scriptDef.MainModule.ImportReference(Weaver.NetworkWriterType)));
         }
 
-        static void WriteCreateWriter(ILProcessor worker)
+        public static void WriteCreateWriter(ILProcessor worker)
         {
             // create writer
             worker.Append(worker.Create(OpCodes.Newobj, Weaver.NetworkWriterCtor));
             worker.Append(worker.Create(OpCodes.Stloc_0));
         }
 
-        static bool WriteArguments(ILProcessor worker, MethodDefinition md, string errString, bool skipFirst)
+        public static bool WriteArguments(ILProcessor worker, MethodDefinition md, string errString, bool skipFirst)
         {
             // write each argument
             short argNum = 1;
@@ -474,7 +473,7 @@ namespace Mirror.Weaver
             m_td.Methods.Add(serialize);
         }
 
-        static int GetChannelId(CustomAttribute ca)
+        public static int GetChannelId(CustomAttribute ca)
         {
             foreach (CustomAttributeNamedArgument customField in ca.Fields)
             {
@@ -779,7 +778,7 @@ namespace Mirror.Weaver
             m_td.Methods.Add(serialize);
         }
 
-        bool ProcessNetworkReaderParameters(MethodDefinition md, ILProcessor worker, bool skipFirst)
+        public static bool ProcessNetworkReaderParameters(TypeDefinition td, MethodDefinition md, ILProcessor worker, bool skipFirst)
         {
             int count = 0;
 
@@ -809,7 +808,7 @@ namespace Mirror.Weaver
                 }
                 else
                 {
-                    Log.Error("ProcessNetworkReaderParameters for " + m_td.Name + ":" + md.Name + " type " + arg.ParameterType + " not supported");
+                    Log.Error("ProcessNetworkReaderParameters for " + td.Name + ":" + md.Name + " type " + arg.ParameterType + " not supported");
                     Weaver.fail = true;
                     return false;
                 }
@@ -817,140 +816,10 @@ namespace Mirror.Weaver
             return true;
         }
 
-        /*
-            // generates code like:
-            protected static void InvokeCmdCmdThrust(NetworkBehaviour obj, NetworkReader reader)
-            {
-                if (!NetworkServer.active)
-                {
-                    return;
-                }
-                ((ShipControl)obj).CmdThrust(reader.ReadSingle(), (int)reader.ReadPackedUInt32());
-            }
-        */
-        MethodDefinition ProcessCommandInvoke(MethodDefinition md)
-        {
-            MethodDefinition cmd = new MethodDefinition(k_CmdPrefix + md.Name, MethodAttributes.Family |
-                    MethodAttributes.Static |
-                    MethodAttributes.HideBySig,
-                    Weaver.voidType);
-
-            ILProcessor cmdWorker = cmd.Body.GetILProcessor();
-            Instruction label = cmdWorker.Create(OpCodes.Nop);
-
-            WriteServerActiveCheck(cmdWorker, md.Name, label, "Command");
-
-            // setup for reader
-            cmdWorker.Append(cmdWorker.Create(OpCodes.Ldarg_0));
-            cmdWorker.Append(cmdWorker.Create(OpCodes.Castclass, m_td));
-
-            if (!ProcessNetworkReaderParameters(md, cmdWorker, false))
-                return null;
-
-            // invoke actual command function
-            cmdWorker.Append(cmdWorker.Create(OpCodes.Callvirt, md));
-            cmdWorker.Append(cmdWorker.Create(OpCodes.Ret));
-
-            AddInvokeParameters(cmd.Parameters);
-
-            return cmd;
-        }
-
-        static void AddInvokeParameters(ICollection<ParameterDefinition> collection)
+        public static void AddInvokeParameters(ICollection<ParameterDefinition> collection)
         {
             collection.Add(new ParameterDefinition("obj", ParameterAttributes.None, Weaver.NetworkBehaviourType2));
             collection.Add(new ParameterDefinition("reader", ParameterAttributes.None, Weaver.scriptDef.MainModule.ImportReference(Weaver.NetworkReaderType)));
-        }
-
-        /*
-            // generates code like:
-            public void CallCmdThrust(float thrusting, int spin)
-            {
-                Debug.LogError("Call Command function CmdThrust");
-                if (!NetworkClient.active)
-                {
-                    Debug.LogError("Command function CmdThrust called on server.");
-                    return;
-                }
-
-                if (isServer)
-                {
-                    // we are ON the server, invoke directly
-                    CmdThrust(thrusting, spin);
-                    return;
-                }
-
-                NetworkWriter networkWriter = new NetworkWriter();
-                networkWriter.Write(thrusting);
-                networkWriter.WritePackedUInt32((uint)spin);
-                base.SendCommandInternal(cmdName, networkWriter, cmdName);
-            }
-        */
-        MethodDefinition ProcessCommandCall(MethodDefinition md, CustomAttribute ca)
-        {
-            MethodDefinition cmd = new MethodDefinition("Call" +  md.Name, MethodAttributes.Public |
-                    MethodAttributes.HideBySig,
-                    Weaver.voidType);
-
-            // add paramters
-            foreach (ParameterDefinition pd in md.Parameters)
-            {
-                cmd.Parameters.Add(new ParameterDefinition(pd.Name, ParameterAttributes.None, pd.ParameterType));
-            }
-
-            ILProcessor cmdWorker = cmd.Body.GetILProcessor();
-            Instruction label = cmdWorker.Create(OpCodes.Nop);
-
-            WriteSetupLocals(cmdWorker);
-
-            if (Weaver.generateLogErrors)
-            {
-                cmdWorker.Append(cmdWorker.Create(OpCodes.Ldstr, "Call Command function " + md.Name));
-                cmdWorker.Append(cmdWorker.Create(OpCodes.Call, Weaver.logErrorReference));
-            }
-
-            WriteClientActiveCheck(cmdWorker, md.Name, label, "Command function");
-
-            // local client check
-            Instruction localClientLabel = cmdWorker.Create(OpCodes.Nop);
-            cmdWorker.Append(cmdWorker.Create(OpCodes.Ldarg_0));
-            cmdWorker.Append(cmdWorker.Create(OpCodes.Call, Weaver.UBehaviourIsServer));
-            cmdWorker.Append(cmdWorker.Create(OpCodes.Brfalse, localClientLabel));
-
-            // call the cmd function directly.
-            cmdWorker.Append(cmdWorker.Create(OpCodes.Ldarg_0));
-            for (int i = 0; i < md.Parameters.Count; i++)
-            {
-                cmdWorker.Append(cmdWorker.Create(OpCodes.Ldarg, i + 1));
-            }
-            cmdWorker.Append(cmdWorker.Create(OpCodes.Call, md));
-            cmdWorker.Append(cmdWorker.Create(OpCodes.Ret));
-            cmdWorker.Append(localClientLabel);
-
-            // NetworkWriter writer = new NetworkWriter();
-            WriteCreateWriter(cmdWorker);
-
-            // write all the arguments that the user passed to the Cmd call
-            if (!WriteArguments(cmdWorker, md, "Command", false))
-                return null;
-
-            var cmdName = md.Name;
-            int index = cmdName.IndexOf(k_CmdPrefix);
-            if (index > -1)
-            {
-                cmdName = cmdName.Substring(k_CmdPrefix.Length);
-            }
-
-            // invoke interal send and return
-            cmdWorker.Append(cmdWorker.Create(OpCodes.Ldarg_0)); // load 'base.' to call the SendCommand function with
-            cmdWorker.Append(cmdWorker.Create(OpCodes.Ldstr, cmdName));
-            cmdWorker.Append(cmdWorker.Create(OpCodes.Ldloc_0)); // writer
-            cmdWorker.Append(cmdWorker.Create(OpCodes.Ldc_I4, GetChannelId(ca)));
-            cmdWorker.Append(cmdWorker.Create(OpCodes.Call, Weaver.sendCommandInternal));
-
-            cmdWorker.Append(cmdWorker.Create(OpCodes.Ret));
-
-            return cmd;
         }
 
         MethodDefinition ProcessTargetRpcInvoke(MethodDefinition md)
@@ -972,7 +841,7 @@ namespace Mirror.Weaver
             //ClientScene.readyconnection
             rpcWorker.Append(rpcWorker.Create(OpCodes.Call, Weaver.ReadyConnectionReference));
 
-            if (!ProcessNetworkReaderParameters(md, rpcWorker, true))
+            if (!ProcessNetworkReaderParameters(m_td, md, rpcWorker, true))
                 return null;
 
             // invoke actual command function
@@ -1000,7 +869,7 @@ namespace Mirror.Weaver
             rpcWorker.Append(rpcWorker.Create(OpCodes.Ldarg_0));
             rpcWorker.Append(rpcWorker.Create(OpCodes.Castclass, m_td));
 
-            if (!ProcessNetworkReaderParameters(md, rpcWorker, false))
+            if (!ProcessNetworkReaderParameters(m_td, md, rpcWorker, false))
                 return null;
 
             // invoke actual command function
@@ -1341,13 +1210,13 @@ namespace Mirror.Weaver
                         names.Add(md.Name);
                         m_Cmds.Add(md);
 
-                        MethodDefinition cmdFunc = ProcessCommandInvoke(md);
+                        MethodDefinition cmdFunc = NetworkBehaviourCommandProcessor.ProcessCommandInvoke(m_td, md);
                         if (cmdFunc != null)
                         {
                             m_CmdInvocationFuncs.Add(cmdFunc);
                         }
 
-                        MethodDefinition cmdCallFunc = ProcessCommandCall(md, ca);
+                        MethodDefinition cmdCallFunc = NetworkBehaviourCommandProcessor.ProcessCommandCall(md, ca);
                         if (cmdCallFunc != null)
                         {
                             m_CmdCallFuncs.Add(cmdCallFunc);
@@ -1493,7 +1362,7 @@ namespace Mirror.Weaver
 
             // read the event arguments
             MethodReference invoke = Weaver.ResolveMethod(eventField.FieldType, "Invoke");
-            if (!ProcessNetworkReaderParameters(invoke.Resolve(), cmdWorker, false))
+            if (!ProcessNetworkReaderParameters(m_td, invoke.Resolve(), cmdWorker, false))
                 return null;
 
             // invoke actual event delegate function
